requirements.txt
beautifulsoup4==4.12.3
lxml==5.1.0
pandas==2.2.0
playwright==1.41.0
python-dotenv==1.0.1
psycopg2-binary
